requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
lxml==5.1.0
pyyaml==6.0.1
jsonschema==4.20.0
Werkzeug==3.0.1
python-docx==1.1.0
